Add a MongoDB web pages registry implementation

+++ crawler/src/main/java/fr/fierdecoder/ajaxcrawlsimulator/crawl/registry/MemoryWebPagesRegistry.java
@@ -25,7 +25,7 @@ public class MemoryWebPagesRegistry implements WebPagesRegistry {
     }
 
     @Override
-    public int getPagesCount() {
+    public long getPagesCount() {
         return pagesByUrl.size();
     }
